requirements.txt
Django>=4.2
djangorestframework
PyPDF2
python-docx
whoosh